Framework for Sql

+++ src/org/fullmetalfalcons/scouting/sql/SqlType.java
@@ -48,7 +48,7 @@ public enum SqlType {
             case SPACE:
                 return null;
             case SLIDER:
-                return INTEGER;
+                return DECIMAL;
             default:
                 return null;
         }
